requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.5.0
av==13.0.0
cachetools==5.5.0
cffi==1.17.1
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
ffmpeg-python==0.2.0
flatbuffers==20181003210633
fonttools==4.54.1
fsspec==2024.10.0
future==1.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.27.0
google-generativeai==0.8.3
google-genai==0.2.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
hailort @ file:///home/pi/hailort-4.18.0-cp311-cp311-linux_aarch64.whl
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
iniconfig==2.0.0
kiwisolver==1.4.7
llvmlite==0.43.0
loguru==0.7.2
markdown-it-py==3.0.0
markdownify==0.13.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
netaddr==1.3.0
netifaces==0.11.0
networkx==3.4.2
numba==0.60.0
numpy==1.23.5
onnxruntime==1.19.2
openai-whisper==20240930
openai==1.60.2
ollama==0.4.7
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
picamera==1.13
picamera2==0.3.21
pidng==4.0.9
piexif==1.1.3
pillow==10.4.0
piper-phonemize==1.1.0
piper-tts==1.2.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
pyTelegramBotAPI==4.23.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-multipart==0.0.12
python-prctl==1.8.1
pyttsx3==2.98
PyYAML==6.0.2
regex==2024.9.11
rich==13.9.2
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
simplejpeg==1.7.6
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.40.0
supervision==0.23.0
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics-thop==2.0.9
uvicorn==0.32.0
uvloop==0.21.0
v4l2-python3==0.3.5
verboselogs==1.7
watchfiles==0.24.0
websockets==13.1
